requirements.txt
Django>=3.0.2,<3.1.0
djangorestframework>=3.11.0,<3.12.0
flake8>=3.7.9,<3.8.0
flake8-import-order>=0.18.1,<0.19
psycopg2>=2.8.4,<2.9.0
Pillow>=7.0.0<7.1.0